% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bisgaard}
\alias{Bisgaard}
\title{Dataset from Bisgaard (2019)}
\format{
A list containing data for government supporters (`gov_data`) and opposition supporters (`opp_data`)
}
\usage{
Bisgaard
}
\description{
Dataset from Bisgaard (2019)
}
\keyword{dataset}
