% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap.R
\name{boot_data}
\alias{boot_data}
\title{Generate Bootstrap data}
\usage{
boot_data(x, seed, data, boot_ind)
}
\arguments{
\item{x}{A}

\item{seed}{A}

\item{data}{A}

\item{boot_ind}{A}
}
\description{
Generate Bootstrap data
}
