% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BroockmanKalla}
\alias{BroockmanKalla}
\title{Dataset from Broockman and Kalla (2018)}
\format{
A list containing the experimental data (`exp_data`) and the target population data (`pop_data`)
}
\usage{
BroockmanKalla
}
\description{
Dataset from Broockman and Kalla (2018)
}
\keyword{dataset}
