% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap.R
\name{gen_bootstrap}
\alias{gen_bootstrap}
\title{Generic Bootstrap function}
\usage{
gen_bootstrap(
  est,
  numCores,
  sims = 1000,
  formula_outcome,
  formula_weights,
  treatment,
  exp_data,
  pop_data,
  weights = NULL,
  pop_weights = NULL,
  boot_ind = NULL,
  seed = 1234
)
}
\arguments{
\item{est}{Function of Estimators}

\item{formula_outcome}{Formula for outcome model}

\item{formula_weights}{Formula for sampling weights}

\item{treatment}{Name of the treatment variable}

\item{exp_data}{Experimental Data}

\item{pop_data}{Population Data}
}
\description{
Generic Bootstrap function
}
