% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Partial Conjunction Test}
\usage{
pct(pvalue, threshold = NULL)
}
\arguments{
\item{pvalue}{a vector of one-sided p-values}

\item{threshold}{A threshold for the partial conjunction test. If NULL, we compute partial conjunction p-values for all thresholds. Default is NULL.}
}
\value{
\code{pct} returns the following values.
 \itemize{
   \item \code{pc}: A table of partial conjunction p-values
   \item \code{tpate}: Estimates of the T-PATE
 }
}
\description{
\code{pct} implements the sign-generalization.
}
\references{
Egami and Hartman. (2020+). Elements of External Validity: Framework, Design, and Analysis
}
