% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls.R
\name{wls}
\alias{wls}
\title{Weighted Least Squares}
\usage{
wls(
  formula_outcome,
  treatment,
  exp_data,
  weights = NULL,
  pop_weights = NULL,
  boot = TRUE,
  sims = 1000,
  boot_ind = NULL,
  numCores = 1,
  seed = 1234
)
}
\description{
Weighted Least Squares
}
