% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls.R
\name{wls_base}
\alias{wls_base}
\title{Basis of Weighted Least Squares}
\usage{
wls_base(
  x,
  formula_outcome,
  formula_weights = NULL,
  treatment,
  exp_data,
  pop_data = NULL,
  weights = NULL,
  pop_weights = NULL,
  boot_ind = NULL,
  seed = 1234
)
}
\arguments{
\item{formula_outcome}{Formula for outcome model}

\item{treatment}{Name of the treatment variable}

\item{exp_data}{Experimental Data}

\item{weights}{Weights should be a vector of numeric values with the same length with `exp_data`}
}
\description{
Basis of Weighted Least Squares
}
