% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpate.R
\name{tpate}
\alias{tpate}
\title{Estimating the Target Population Average Treatment Effect (TPATE)}
\usage{
tpate(
  formula_outcome,
  formula_weights,
  treatment,
  exp_data,
  pop_data,
  est_type = "wls",
  weights_type = "calibration",
  weights_max = Inf,
  pop_weights = NULL,
  boot = TRUE,
  sims = 1000,
  numCores = NULL,
  seed = 1234,
  ...
)
}
\arguments{
\item{formula_outcome}{Formula for outcome model}

\item{formula_weights}{Formula for sampling weights}

\item{treatment}{Name of the treatment variable}

\item{exp_data}{Experimental Data}

\item{pop_data}{Population Data}

\item{estimator_type}{Whether we use a paired-choice conjoint design}
}
\value{
\code{model_pAMCE} returns an object of \code{pAMCE} class.
 \itemize{
   \item \code{AMCE}: Estimates of the pAMCE for all factors.
   \item \code{boot_AMCE}: Estimates of the pAMCE for all factors in each bootstrap sample.
   \item \code{boot_coef}: Estimates of coefficients for the linear probability model in each bootstrap sample.
   \item \code{approach}: "model_based"
   \item \code{input}: Input into the function.
   \item \code{...}: Values for internal use.
 }
}
\description{
\code{tpate} implements the effect-generalization.
}
\references{
Egami and Hartman. (2020+). Elements of External Validity: Framework, Design, and Analysis
}
