% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpate.R
\name{tpate}
\alias{tpate}
\title{Estimating the Target Population Average Treatment Effect (TPATE)}
\usage{
tpate(
  formula_outcome,
  formula_weights,
  treatment,
  exp_data,
  pop_data,
  est_type = "wls",
  weights_type = "calibration",
  weights_max = Inf,
  boot = TRUE,
  id_cluster = NULL,
  sims = 1000,
  numCores = NULL,
  seed = 1234,
  compute_sate = TRUE,
  ...
)
}
\arguments{
\item{formula_outcome}{Formula for outcome model (should include treatment)}

\item{formula_weights}{Formula for sampling weights}

\item{treatment}{Name of the treatment variable}

\item{exp_data}{Experimental Data}

\item{pop_data}{Population Data}

\item{est_type}{Estimator Type; `ipw`, `outcome-ols`, `outcome-bart`, `dr-ols`, `dr-bart`, or `wls-proj`}

\item{weights_type}{Weights Type; `logit`, `calibration`}

\item{weights_max}{Default is `Inf`}

\item{boot}{Whether you do bootstrap (`TRUE` or `FALSE`)}

\item{id_cluster}{Identifies for cluster bootstrap}

\item{sims}{The number of simulations}

\item{numCores}{The number of cores we use}

\item{seed}{seed (default = `1234`)}

\item{compute_sate}{whether we compute SATE}
}
\value{
\code{tpate} returns the following values.
 \itemize{
   \item \code{sate}: Estimates of the SATE
   \item \code{tpate}: Estimates of the T-PATE
 }
}
\description{
\code{tpate} implements the effect-generalization.
}
\references{
Egami and Hartman. (2020+). Elements of External Validity: Framework, Design, and Analysis
}
