% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpate.R
\name{tpate}
\alias{tpate}
\title{Estimating the Target Population Average Treatment Effect (TPATE)}
\usage{
tpate(
  outcome,
  treatment,
  covariates,
  covariates_exp = NULL,
  exp_data,
  pop_data,
  est_type = "wls",
  weights_type = "calibration",
  weights_max = Inf,
  boot = TRUE,
  id_cluster = NULL,
  sims = 500,
  numCores = NULL,
  seed = 1234,
  compute_sate = FALSE
)
}
\arguments{
\item{outcome}{Mame of the outcome variable}

\item{treatment}{Name of the treatment variable}

\item{covariates}{Names of covariates we adjust for.}

\item{covariates_exp}{(Optional) Names of covariates measured only in the experimental data. These covariates are used only in `wls`}

\item{exp_data}{Experimental data. This should be `data.frame`}

\item{pop_data}{Population data. This should be `data.frame`}

\item{est_type}{Estimator Type. Should be one of the six types (`ipw`, `wls`, `outcome-ols`, `outcome-bart`, `dr-ols`, `dr-bart`). Weighting-based estimators are `ipw` and `wls`. Outcome-based estimators are `outcome-ols` and `outcome-bart.` Doubly-robust estimators are `dr-ols` and `dr-bart`.}

\item{weights_type}{Weights Type. Should be one of the two types (`calibration`, `logit`). We recommend `calibration` for its stability in many applications.}

\item{weights_max}{Maximum weights. Default is `Inf`.}

\item{boot}{Logical. Whether we use bootstrap (`TRUE` or `FALSE`). For doubly robust estimators (`dr_ols` and `dr-bart`), we always use bootstrap (`boot = TRUE`).}

\item{id_cluster}{Identifies for cluster bootstrap. The length should be the same as the number of rows of `exp_data`.}

\item{sims}{The number of bootstrap replications (when `boot = TRUE)`. The number of quasi-bayesian posterior samples (when `boot = FALSE`).}

\item{numCores}{The number of cores we use for parallel computing.}

\item{seed}{seed (default = `1234`)}

\item{compute_sate}{Logical. Whether we compute the SATE within the package.}
}
\value{
\code{tpate} returns the following values.
 \itemize{
   \item \code{sate}: Estimates of the SATE
   \item \code{tpate}: Estimates of the T-PATE
   \item \code{ipw_weights}: Estimated weights (when weighting-based or doubly robust estimators are used)
 }
}
\description{
\code{tpate} implements the effect-generalization.
}
\references{
Egami and Hartman. (2022+). Elements of External Validity: Framework, Design, and Analysis.
}
