% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls.R
\name{wls}
\alias{wls}
\title{Weighted Least Squares}
\usage{
wls(
  formula_outcome,
  treatment,
  exp_data,
  weights = NULL,
  pop_weights = NULL,
  boot = TRUE,
  sims = 1000,
  boot_ind = NULL,
  numCores = 1,
  seed = 1234
)
}
\arguments{
\item{formula_outcome}{Formula for outcome model}

\item{treatment}{Name of the treatment variable}

\item{exp_data}{Experimental Data}

\item{weights}{Weights should be a vector of numeric values with the same length with `exp_data`}
}
\description{
Weighted Least Squares
}
